[Feat] DreamLog #121 응원로그 저장방식 변경 & 메인뷰에 데이터 띄우기
