+++ mc2-DreamLog/mc2-DreamLog/Global/Extension/ViewExtension.swift
@@ -45,5 +45,23 @@ extension View {
         controller.view.removeFromSuperview()
         return image
     }
+    
+    func dateConverter(inputDateString: String) -> String {
+        
+        var dateString = inputDateString
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
+        let date = dateFormatter.date(from: dateString)
+        
+        dateFormatter.dateFormat = "yyyy-MM-dd"
+        dateString = dateFormatter.string(from: date!)
+        
+        let calendar = Calendar.current
+        let updatedDate = calendar.date(byAdding: .hour, value: 9, to: date!)
+        
+        let updatedDateString = dateFormatter.string(from: updatedDate!)
+        
+        return updatedDateString
+    }
 }
