+++ mc2-DreamLog/mc2-DreamLog/View/CheerLogView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import CoreData
 
 struct CheerLogView: View {
     var body: some View {
@@ -31,107 +30,42 @@ struct CheerLogView: View {
 
 struct CheerList: View {
     
-    @StateObject var model = dataModel()
+    @State var cheerList: [CheerLogModel] = []
     
     var body: some View {
         VStack {
             
             List{
                 
-                ForEach(model.data, id: \.objectID) { obj in
-                    //Extracting data from obj
-                    let (cheer, writtenDate) = model.getValue(obj: obj)
-                    
+                ForEach(Array(cheerList.enumerated()), id: \.1) { (index, cheerData) in
                     VStack(alignment: .leading) {
-                        Text(cheer)
+                        Text(cheerData.cheer)
                             .multilineTextAlignment(.leading)
                             .padding(.bottom, 8)
-                        Text(writtenDate)
+                        
+                        Text(dateConverter(inputDateString: cheerData.time))
                             .font(.system(size: 14, weight: .regular))
                             .foregroundColor(.gray)
                     }
                 }
-                .onDelete(perform: model.deleteData(indexSet:))
+                .onDelete { indexSet in
+                    let indexes = indexSet.map { $0 }
+                    for index in indexes {
+                        let cheerData = cheerList[index]
+                        let i = cheerList.firstIndex(of: cheerData)!
+                        DBHelper.shared.deleteCheerLogData(id: cheerList[i].id)
+                        cheerList = DBHelper.shared.readCheerLogData()
+                    }
+                }
+                
             }
             .listStyle(InsetGroupedListStyle())
             .onAppear {
-                model.readData()
+                self.cheerList = DBHelper.shared.readCheerLogData()
             }
         }
     }
 }
-
-
-
-class dataModel: ObservableObject {
-    @Published var data: [NSManagedObject] = []
-    @Published var cheerText = ""
-    @Published var writtenDateText = ""
-    let context = persistentContainer.viewContext
-    
-    init() {
-        readData()
-    }
-    
-    func readData() {
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CheerData")
-        
-        do {
-            let results = try context.fetch(request)
-            
-            self.data = results as! [NSManagedObject]
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
-    
-    func writeData() {
-        let entity = NSEntityDescription.insertNewObject(forEntityName: "CheerData", into: context)
-        
-        entity.setValue(cheerText, forKey: "cheer")
-        entity.setValue(writtenDateText, forKey: "writtenDate")
-        
-        do {
-            try context.save()
-            self.data.append(entity)
-            
-            cheerText = ""
-            writtenDateText = ""
-            
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
-    
-    func deleteData(indexSet: IndexSet) {
-        
-        for index in indexSet {
-            
-            do {
-                let obj = data[index]
-                context.delete(obj)
-            
-                try context.save()
-                
-                let index = data.firstIndex(of: obj)
-                data.remove(at: index!)
-            } catch {
-                print(error.localizedDescription)
-            }
-            
-        }
-    }
-    
-    func getValue(obj: NSManagedObject) -> (cheer: String, writtenDate: String) {
-        let cheer = obj.value(forKey: "cheer") as! String
-        let writtenDate = obj.value(forKey: "writtenDate") as! String
-        
-        return (cheer, writtenDate)
-    }
-}
-
-
-
 
 struct CheerLogView_Previews: PreviewProvider {
     static var previews: some View {
