+++ mc2-DreamLog/mc2-DreamLog/View/DreamBoardView.swift
@@ -10,10 +10,10 @@ import SwiftUI
 struct DreamBoardView: View {
     @EnvironmentObject var data: TutorialBoardElement
     @State var isDone = false
-    @State var text = ""
+    @State var cheertext = ""
+    @State var alerttext = ""
     @State private var showingAlert: Bool = false
     @State private var confirmAlert: Bool = false
-    @StateObject var cheerModel = dataModel()
     @State private var boardImage: UIImage = Tab1Model.instance.image ?? UIImage(named: "BoardDummy")!
     
     @State private var showDDayCalendar = false
@@ -33,15 +33,13 @@ struct DreamBoardView: View {
             let width = geo.size.width
             let height = geo.size.height
             
-            
-            
             VStack {
                 
                 VStack(spacing: 0) {
                     
                     Image(uiImage: boardImage)
                     
-                    Text(text == "" ? "스스로를 위한 응원을 작성해보세요" : text)
+                    Text(cheertext == "" ? "스스로를 위한 응원을 작성해보세요" : cheertext)
                         .grayText(fontSize: 22)
                         .fontWeight(.semibold)
                         .frame(width: abs(width), height: 40, alignment: .center)
@@ -103,28 +101,30 @@ struct DreamBoardView: View {
                                 .foregroundColor(.textGreen)
                         }
                         .alert("나에게 주는 응원 한마디를\n작성해주세요", isPresented: $showingAlert, actions: {
-                            TextField("응원의 한 마디를 작성해보아요", text: $cheerModel.cheerText)
+                            TextField("응원의 한 마디를 작성해보아요", text: $alerttext)
                             
                             Button("완료", action: {
                                 confirmAlert = true
                             })
-                            Button("취소", role: .cancel, action: {})
+                            Button("취소", role: .cancel, action: {
+                                alerttext = ""
+                            })
                         })
                         .alert(isPresented: $confirmAlert, content: {
-                            Alert(title: Text("\(cheerModel.cheerText)으로\n응원을 추가하시겠어요?"),
+                            Alert(title: Text("\(alerttext)으로\n응원을 추가하시겠어요?"),
                                   message: Text("작성하신 응원은 위젯에 표시됩니다."),
                                   primaryButton: .default(Text("확인"), action: {
-                                text = cheerModel.cheerText
-                                cheerModel.writtenDateText = getCurrentDate()
-                                cheerModel.writeData() // 첫 번째 액션
-                                print(getCurrentDate())
-                                print($cheerModel.cheerText)
-                                
-                            }),
-                                  secondaryButton: .cancel(Text("취소"), action: {
-                                // 액션 없음
+                                // 데이터 처리
+                                DBHelper.shared.createCheerLogTable()
+                                DBHelper.shared.insertCheerLogData(alerttext)
+                                cheertext = DBHelper.shared.readCheerLogDataOne().cheer
+                            }), secondaryButton: .cancel(Text("취소"), action: {
+                                cheertext = DBHelper.shared.readCheerLogDataOne().cheer
+                                alerttext = ""
                             }))
+                            
                         })
+                        
                     }
                 }
                 .padding(.horizontal, 16)
@@ -136,14 +136,10 @@ struct DreamBoardView: View {
                 .shadow(color: Color.shadowGray, radius: 2, x: 0, y: 2)
             }
             .onAppear {
-                // 이것도 샘플 이미지로 바꿔주는 작업
-                self.boardImage = imageFileManager.getSavedImage(named: DBHelper.shared.readDreamLogDataOne().imagePath) ?? UIImage(named: "sticker_check")!
-            }
-            .onAppear {
                 getDDayDate()
+                self.boardImage = imageFileManager.getSavedImage(named: DBHelper.shared.readDreamLogDataOne().imagePath) ?? UIImage(named: "sticker_check")!
+                cheertext = DBHelper.shared.readCheerLogDataOne().cheer
             }
-            
-            
         }
     }
     
@@ -180,4 +176,6 @@ struct MainTab1View_Previews: PreviewProvider {
         }
     }
 }
+
+
